requirements.txt
fastapi==0.125.0
httpx==0.28.1
joblib==1.5.2
joblib==1.5.3
matplotlib==3.10.8
nltk==3.8.1
nltk==3.9.2
numpy==2.3.5
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.2.1
scikit_learn==1.3.0
seaborn==0.13.2
